+++ package.json
@@ -13,6 +13,10 @@
   "devDependencies": {
     "@babel/cli": "7.11.6",
     "@babel/core": "7.11.6",
+    "@types/lodash": "^4.14.187",
+    "@types/node": "^18.11.9",
+    "@typescript-eslint/eslint-plugin": "^5.42.0",
+    "@typescript-eslint/parser": "^5.42.0",
     "babel-eslint": "10.1.0",
     "babel-preset-gatsby-package": "0.5.3",
     "cross-env": "7.0.2",
@@ -23,7 +27,11 @@
     "eslint-plugin-jsx-a11y": "6.3.1",
     "eslint-plugin-prettier": "3.1.4",
     "eslint-plugin-react": "7.20.6",
-    "prettier": "2.1.2"
+    "gatsby": "^4.24.7",
+    "gatsby-source-filesystem": "^4.24.0",
+    "lodash": "^4.17.21",
+    "prettier": "2.1.2",
+    "typescript": "^4.8.4"
   },
   "homepage": "https://github.com/oorestisime/gatsby-source-instagram",
   "keywords": [
@@ -32,8 +40,8 @@
   ],
   "license": "MIT",
   "peerDependencies": {
-    "gatsby-source-filesystem": "^2.0.27",
-    "gatsby": ">2.0.0-alpha"
+    "gatsby": ">2.0.0-alpha",
+    "gatsby-source-filesystem": "^2.0.27"
   },
   "repository": "https://github.com/oorestisime/gatsby-source-instagram",
   "scripts": {
